Fixes to ArrayInit devirtualization
Summary: This is a diff on top of a revert of the revert of
devirtualizing ArrayInit for mixed arrays.  (It broke the build
because in another diff I was changing ::SetStr and ::SetInt to take
Cells.)  We don't need the KindOfUninit check outside right now, but
if it's in the noise in perflab let's do it this way (we can
eventually remove it on the inside).

Reviewed By: eletuchy

Subscribers: hphp-diffslists, ps

Differential Revision: D1333696