Fix inlining correctness with regard to stack overflows
Summary: The code in shouldIRInline that tries to count how deep we're
inlining apparently doesn't actually run when the RegionIter is a
TraceletIter.  Code that limits this is required for correctness
because we omit stack overflow checks on inlined calls.

For this diff I'm not changing the logic in shouldIRInline, except to
add comments that it's not something you can use for anything you need
to check for correctnes because it doesn't happen in the tracelet
mode.

Reviewed By: @edwinsmith

Differential Revision: D1335581