Make DefInlineSP non-essential, clarify how frame use counts work in dce
Summary: This function was changed to have 'use counts' that aren't
use counts.  What's going on is some of the transitive uses are
counted as uses and some aren't, which doesn't really make sense to me
(or if it does, it should have a doc comment explaining what is going
on) .... Lack of clarity here seems to be why we had the weird buggy
removal of DefInlineFPs without removing the nested SpillFrames in the
recursive inlining test (which was worked around by D1326900).

This diff reverts to something more like the original way it worked:
the stack definition after the FP definition is back to a
non-essential instruction (it used to just be another ReDefSP), so we
can just do normal mark/sweep-style dce to eliminate it.  All uses of
the stack in the callee will end up keeping a use on the frame
indirectly via this chain.  DCE will only eliminate the DefInlineSP if
it's unused, and therefore the DefInlineFP won't go away if the stack
is used, so the problem we had with recurse.php isn't a problem
anymore.  Also I'm pretty sure PassFP and PassSP are not necessary,
but I haven't removed them yet.  (I'm going to try to rewrite this as
a separate pass today to make it support eliminating frames wrapping
Call instructions also.)

Also this makes StLocNT not keep frames alive.  (I observed them
preventing ActRec removal in a test case.)

Reviewed By: @ottoni

Differential Revision: D1335948