Clean up ActRec elision, partial implementation for Call as weak frame uses
Summary: This removes PassFP, PassSP, DefInlineSP, the list of frames
in ReDefSP's extra data, the old code that tried to eliminate frames
around calls, and simplifies the DecRef stack depth fixups to just
move it down to the known safe depth.  Also split optimizeActRecs into
a some subroutines.  I was planning to move this to a different file
than dce, but it kinda seems like it's actually nice for it to share
the normal marking and sweep steps for now, so we might as well keep
it here until/unless it grows bigger (or e.g. if we need it to iterate
to remove multiple frames in nested calls some day, which we might not
want to happen on every dce pass).

Reviewed By: @swtaarrs

Differential Revision: D1343427