CreateAFWH consumes $this
Summary: This was not accounted for properly in ref-counting optimizations.

A LdThis instuction counts as creating a reference for $this (even
if there is no IncRef $this), and it must be explicitly consumed. On a
normal control flow return path, RetC consumes $this with a DecRefThis. But,
if we return with a suspended Awaitable, we were not counting that the
CreateAFWH consumed $this.

This bug was not triggered so far because we would always break
the tracelet before an Await instruction (usually the control
flow is like IsType<Null>; JmpNZ; Await, and we break the tracelet
at JmpNZ). But when I enabled tracing through JmpZ/JmpNZ, it allowed
a previous LdThis reference to remain alive at the point of Await.

Reviewed By: jan

Subscribers: hphp-diffslists, ps

Differential Revision: D1336287