Fix Xenon crash
Summary: When WaitableWaitHandle::getdependencystack is called, there
could be async functions on the stack which are in the middle of returning.
In particular, their $this might have been destroyed, and then we have
started executing a local variable's __destruct, which then ends up
calling WaitableWaitHandle::getdependencystack (possibly through many
levels of calls).

This typically happens due to Xenon being called from onFunctionEnter,
or onFunctionExit. There was a previous attempt to fix these crashes
by nulling out $this in the ActRec passed to onFunctionExit; but as
described above (and as happens in the new test case), there's no
guarantee that the problem is happening in the top frame.

Reviewed By: drussi

Subscribers: hphp-diffslists, jan, ps, bill

Differential Revision: D1336321

Tasks: 4352949