s/Zend PHP/PHP5/g
Summary: Zend didn't write it and there are members of the comunity which prefer this name. We don't really want to take a stance either way, so we're going with PHP5 as the name for the reference implementaiton of php.

I removed all the extra 5s from `PHP5 5.` which this introduced.

Reviewed By: @JoelMarcey

Differential Revision: D1336440