Stop making a copy in EmptyArray::Copy
Summary: Ideally we shouldn't need to copy the empty array.

This broke ArrayUtil::Walk because it started calling
EmptyArray::AdvanceMArrayIter. Switching that to return false instead
of not_reached() fixed things locally for me.

Also broke GenArrayWaitHandle::create, which was asserting that it
would only get Packed or Mixed arrays.

Reviewed By: @jano

Differential Revision: D1299321