Truncate very long socket paths
Summary: Apparently the max length for these is much shorter than the
max length for a path? In any event I get a reproducible failure when
using the test case in the linked issue. Let's enforce a max -- the
number 50 was chosen pretty arbitrarily, not sure what it really should
be. (Didn't want to digest universally since this could harm
debuggability in the common case.)

Also deleted some dead code.

Fixes #2149

Reviewed By: @pikatchu

Differential Revision: D1322841