Add support for multiple cookies with same name, Max-Age, take 2
Summary: PHP5 does not dedupe cookies. This is in conflict with RFC6265,
which states that web servers should not send two set-cookie headers
with the same cookie-name in the same response. However, existing
frameworks such as Wordpress rely on using the same cookie name with
different paths to function properly. This diff takes the compromise
approach. We will dedupe on the triplet of name-domain-path. From an
end-user impact, this means that we take over the job of deduping from
the client, avoiding undefined behavior for a well defined last-one-wins
policy.

This is motivated by D1316339, which was reverted due to breaking
internal tools.

Reviewed By: @ptarjan

Differential Revision: D1338710