Add camel caps search support to hh_client search
Summary: This adds a way to search for a long class name faster if you know the capital letters in it.

For example, you could search for FooBarClass by typing 'fbc' instead of the full name.

This was supported in monocle and I don't want to drop any features so I've added it to hh_client search.

Reviewed By: @jwatzman

Differential Revision: D1339161