Add "hhvm.server.ip" ini setting, matching "Server.IP" HDF setting
For exemple, this allows to specify which IP address the FastCGI server
will listen on. On a dual-IP-stack system (IPv4+IPv6), HHVM would listen
on "::" by default (IPv6). With this setting, one can make it listen
to "0.0.0.0" instead (IPv4) using a php.ini file, instead of an HDF
configuration file.