Fix abstract constructors
Summary: No need to call `parent::__construct` if it's abstract (and, of
course, you actually can't call it :)), so flip off that bit. In order
to make this work properly with interfaces, fix a bug where we were
considering all interface methods to be implicitly abstract except for
the constructor.

I'm not a huge fan of abstract constructors, but there are cases where
they make sense, even more so when/if we get first-class classes or
something similar.

Fixes #2158

Reviewed By: @pikatchu

Differential Revision: D1340379