Spill stack before potential side exits in hottrace regions
Summary: For blocks that have multiple successors in the TransCFG, mark them as
potential side exits in the region.  In translateRegion, force a spill
stack after such blocks.  This avoids any code in the exit block,
which then allows the jump in the main code to be patched and avoids
going to astubs.

Reviewed By: @bertmaher

Differential Revision: D1335866