print the actual command for hhbbc mode
Summary: *shakes fist at @jdelong*. I just spent an hour trying to figure out why the command it printed out wasn't failing. It was because the `Repo.Authoritative` was being added specially after, and wasn't being shown to the user that it was needed.

Any ideas how to not have this happen again?

Reviewed By: @JoelMarcey

Differential Revision: D1342039