Make recordBCInstr less broken
Summary: It was trying to figure out where the code was by
looking at mcg->code.stubs(), which is wrong. The callsites
should always know the correct value though, so let them decide.

It turns out this is mostly dead anyway - only a handful of things
get recorded, and they're all in "a".

Closes #2712

Reviewed By: aravind

Differential Revision: D1342628