Pass ObjectData* instead of const Object& to HNI methods, remove null checks
Summary: The accesses to -> on Object will do a null check, and most
of the code also had explicit null checks and instanceof checks.  Also
at runtime this means passing a pointer to a pointer, which the
compiler may repeatedly have to reload if you make calls to other
translation units because it won't know that the ObjectData* in the
Object didn't change.

Reviewed By: @sgolemon

Differential Revision: D1347086