Try to inline singleton pattern functions
Summary: We see a lot of code roughly of the form

  function getSingletonFoo() {
    static $instance = null;
    if ($instance === null) {
      // Do some stuff and set $instance.
    }
    return $instance;
  }

Let's optimistically assume that such patterns are in fact singleton
patterns (as opposed to, say, some means of inter-method communication
to avoid threading arguments) and try to inline these checks.  The first
time we hit each callsite, we'll side-exit to do the call, and
subsequent calls will simply return the value at $instance, assuming the
checks still pass.

We also introduce a TransFlags struct which we thread through BindJmp
requests and into the translation routines so that we turn the inlining
off in the stub.

Presently, this inlining strategy is restricted to non-resumable
functions with no arguments passed and which don't return refs.

Reviewed By: bertrand

Differential Revision: D1344602