hphp/runtime/server/upload.cpp: fix "=" vs "==" assert typo
Summary: I caught a similar typo in a review an hour ago, so went searching
through all of fbcode for other occurrences of this type of error.

Coincidentally, I see that D1311333 has just gone in,
and it made some of the same changes I had made, but this
one still remains.

Reviewed By: @markw65

Differential Revision: D1342164