Enable HHIR guard relaxation for Optimize translations
Summary: This was almost working apparently.  This diff fixes a couple of bugs
and got this working again.

Profile translations are not relaxed, and thus types are never
specialized.  We can try that separately, but I'm thinking that it
might be best to specialize types based on profiling instead (as done
for ArrayGet).

Reviewed By: @swtaarrs

Differential Revision: D1342724