implement $php_errormsg
Summary: @fredemmott did all the heavy lifting with `$http_response_header` I just copied that diff.

This doesn't match PHP5 perfectly in that it doesn't define the variable if the ini setting is off and the variable isn't defined at the start of the request. I'm assuming fred did that for perf so I'm just going to match it for now and fix both together if it is needed.

Closes #1164

Reviewed By: @markw65

Differential Revision: D1338752