Fix XLS intervals representing caller-saved registers for native calls
Summary: Intervals for caller-saved registers were ending past the native call,
which then resulted in the these registers interfering with the dest
registers of the native calls.  As a result, the dest registers of
native-call instructions were always assigned to callee-saved
registers, which had 2 bad side-effects: (1) it increased the pressure
on these registers; and (2) it caused unnecessary register moves from
the function-return regs (rax/rdx) to the assigned callee-saved regs.

Reviewed By: @edwinsmith

Differential Revision: D1343728