Remove Redis::_eval* methods, apply minor fixes and cleanups
Summary: The _eval method was named so because eval is a reserved word in php, _evalSha was introduced for the sake of consistency with _eval. I believe that we do not need this methods anymore since we could use Redis::eval() (or Redis::evaluate()), thanks to method aliases.

Closes #2627

Reviewed By: @ptarjan

Differential Revision: D1345010

Pulled By: @JoelMarcey