Avoid unnecessary writes to AsyncFunctionWaitHandle::m_child
Summary: AsyncFunctionWaitHandle::m_child contains valid pointer only in
STATE_SCHEDULED or STATE_BLOCKED. This pointer was checked for null only
in the constructor when were were in one of these 2 states. Kill this
check as it always succeeds.

Do not set m_child to null when transitioning to STATE_SUCCEEDED ot
STATE_FAILED, or when constructing the wait handle -- the value is set
in initialize().

Reviewed By: @jdelong

Differential Revision: D1346943