Create a CodeGenFixups class
Summary: Create a CodeGenFixups class, to hold temporary information that will eventually be put in global tables. Put m_pendingFixups, m_pendingCatchTraces and a newly created m_pendingJmpTransIDs there.

The goal is to have a single place to hold information that may need to be fixed up if we modify the generated code before committing it (eg by relocating it, or by peephole optimizations). Also, these are things that must not go into the global tables if we end up rolling back.

Reviewed By: @ottoni

Differential Revision: D1347532