Prepare child for establishing dependency from within Await opcode
Summary: When awaiting a WaitHandle from resumed async function, the child
WaitHandle must be prepared for establishing dependency:

- if the child WaitHandle is not associated with the currently running scheduler, it must be imported; this will throw if cross-scheduler cycle is formed
- a cycle check must be performed

This preparation is currently done after await returns back control to
the scheduler. If an exception occurs, the failing await is resumed and
exception rethrown.

Let's do this from within the Await opcode so that in this unlikely
case, unnecessary reentry is saved. What is perhaps more important, this
brings us closer to avoiding unnecessary scheduler reentries.

Reviewed By: @billf

Differential Revision: D1347381