Refactor parent wakeup
Summary: Once a WaitHandle is finished, it needs to wake up all its parents.
Obtain pointer to the parents before the result is set (so that we can
union m_resultOrException with m_firstParent+m_nextParent).

We can't move m_firstParent to the union yet because RetC opcode sets
result before giving the control back to the scheduler.

Reviewed By: @billf

Differential Revision: D1348077