Get rid of SmashTarget
Summary: The SmashTarget is a short lived, malloced data block that gets passed
to handlePrimeCacheInit, telling it how to smash the call. The information it
contains is the address of the movq instruction that needs smashing, and the
return address of the handlePrimeCacheInit method.

We can find the return address by inspecting the stack, and then we can
pass in the offset from that to the movq instead of passing in the address
of a SmashTarget.

In addition to avoiding the malloced data, this has the nice effect of making
all the code position independent too.

Reviewed By: @mxw

Differential Revision: D1348201