Workaround bug and swap libedit/libreadline order
The build system is a bit naive and picks linker arguments based on the
presence of libraries and not based on the library choice that was made
earlier which in turn, was based on the availability of headers.

This has the effect that if you have both libedit and libreadline
installed on the system but only libedit-dev, hhvm will get built with
libedit headers but link will be attempted against libreadline which,
obviously fails.

Since we want to prefer libedit over libreadline in the Debian package
for licensing reasons, work around this bug by ordering libedit in this
	first.  This is, of course, a workaround and not a fix as the
	opposite could happen as well.