Convert null_object/resource references
Summary: We've been using null_* when we really want to just return an
empty/new Object/Resource/String/Array. This is in preparation
for a codemod that will convert these sites to null_foo_value but gets
the short-term win. Handle Variant in a separate diff because it of the
semantics.

Reviewed By: @markw65

Differential Revision: D1344569