Exit immediately on SIGTERM in server mode
In server mode, the HHVM server will now always terminate immediately
if SIGTERM is received and a (default off) Server.KillOnSIGTERM
hdf configuration setting is on. Previously, SIGTERM signals would
initiate a graceful shutdown of the server, which runs against
precedent in unix servers. This previous behavior is maintained
when the Server.KillOnSIGTERM option is false, which is by default.