Teach HHVM's JIT about count($value)
Summary: The idea is to fast path accessing the size of an array or a
collection. Collections always have their size at
FAST_COLLECTION_SIZE_OFFSET, and arrays (that are not NVTW) have their
size at ArrayData::offsetofSize(). For collections we can just do a
load, but for arrays of unknown kind we have to do an if/else in
codegen on whether to do the load directly or to call out to
ArrayData::size().

Reviewed By: @swtaarrs

Differential Revision: D1335482