setup the machinery for ini_get to work
Summary: This introduces a new api, `Config::Bind` which takes the param by reference instead of returning the value. This way we can remember the location for future `ini_get` calls.

This diff doesn't move any of the api users over, I want to do that in smaller chunks so they are easier to revert if they cause issues since they'll be read pipeline chanes.

Reviewed By: @fredemmott

Differential Revision: D1348849