Eagerly shrink branches
Summary: Eagerly shrink branches as we relocate, as long as we can
prove that the target will be in range.

I decided against trying the optimal solution for now, since its
complicated by the instruction alignment rules. All the necessary
information is available, but I think this will get nearly all the
benefits. Im open to persuasion if anyone thinks differently.

Reviewed By: @swtaarrs

Differential Revision: D1357920