Revert "Implement type hierarchy for interfaces"
Summary: This reverts commit fa23cfbca5ba3ecb3eb60ef56029c379c2722927.
The Class::subtypeOf function isn't correct, and the logic is a bit
complicated so I'd rather revert for now than try to figure out how to
do it better.  (aravind: if you want to look, I think the issue is
that it's not clear what a non-empty interface list in a res::Class is
supposed to mean.)

Reviewed By: @dariorussi

Differential Revision: D1359861