Read hhconfig to determine if we should use embedded hhi's
Summary: This is separate from D1353049 so that it can be separately
reverted. We need a transition plan for FB www, and this is simple. This
isn't such a huge deal for open-source users -- they can just delete
their hhi directories as part of their hh_server upgrade. For more
involved upgrades, we plan to go this path, except actually go an entire
OSS release with the config option present, and all sorts of automatic
upgrade stuff, but we don't have the machinery for that quite yet.

Reviewed By: @elgenie

Differential Revision: D1364920