Allow vtable-style dispatch on abstract instance methods
Summary: This was disabled, going back to the initial implementation
in tx64.  Abstract methods have method slots that behave the same when
overriden as normal methods, so maybe this dates back to before we had
bytecode bodies for the "base" abstract methods (to implement "pure
virtual called" errors)?  This diff only enables it for non-static
methods.

Reviewed By: @ottoni

Differential Revision: D1366913