Make array<some-supertype> usable in strict
Summary: Now that we allow arrays in strict, as long as they are fully
annotated, it makes sense to relax the "heterogeneous array" error,
which prevents you from writing `array<mixed>` or even
`array<someinterface>` (see test case) which are both useful things to
do, particularly the latter.

This does leave some cruft with regards to how this is handled in
partial mode due to www errors and upgrade/migration concerns, see code
comment.

Reviewed By: @elgenie

Differential Revision: D1368101