remove mtime and prefetch
Summary: A long long time ago, the hack server was using message passing to implement the shared heap.
Back then, it was faster to batch the messages. Because of that, we were keeping the set of
classes and functions used by a node in the previous run to prefetch them the next time.
All of this is completely useless today, since we use shared memory, batching is not helping
what so ever!

Reviewed By: @elgenie

Differential Revision: D1365410