Clean up some ir state merging issues
Summary: I traced a test failure on D1365041 to the fact that we weren't always
propagating state between blocks when we needed to. The biggest issue is that
the done_block in IRBuilder::cond wasn't empty when we appended it, so we
weren't loading up its saved state. I added an assert that we can't enter a
block with predecessors unless it also has saved incoming state, which caught a
bigger issue with how IRBuilder::getKnownTypes() works. This has always been
pretty sketchy, so I fixed it properly by making it its own, full pass over the
IR after all optimizations are finished.

removeGuard() in guard-relaxation.cpp has been eliminated because it was
causing more state propagation issues and hasn't been needed ever since we made
the Check/AssertType opcodes intersect their input and type params.

Reviewed By: @ottoni

Differential Revision: D1369383