Fix hhir guard relaxation for specialized types
Summary: DataTypeSpecialized by itself isn't a strong enough constraint: we
also need to know which specific class the caller wants or if they want an
ArrayKind. This diff adds support for storing this information in
TypeConstraint and all the associated machinery.

The problem was using DataTypeSpecialized to constraint a CheckType<Packed> on
an array that already had a RepoAuthType::Array in its type: the source type
was already specialized so we thought it was safe to not constrain the
CheckType, leading to a not necessarily Packed array being given to
LdPackedArrayElem. To prevent this from happening undetected in the future, I
added support for array kinds to the IR opcode table.

Reviewed By: @jdelong

Differential Revision: D1369432