Fix a race condition in hh_add.
Summary: Fixed a documented race condition in hh_add in the Hack lockfree hashtable. If two threads try to insert the same data, and both try to grab the same slot, the one that fails could return before the one that won writes the (char*)1 sigil indicating that it is filling in the value. The solution is to still try to write the value in in this case.

Reviewed By: @gabelevi

Differential Revision: D1361974