Improve tracing in global DCE, and add some more comments in analyze_dce
Summary: It was confusing to both me and Owen that allLive gets every
bit set during the analyze pass, but this is actually correct (we
don't know whether anything is live or not yet---so the visitor needs
to conservatively assume it could be live).  We compute the gen and
kill sets during this pass, but liveOut is initialized to the empty
set for each block in the next part when do the fixed point iteration.

Reviewed By: @alexmalyshev

Differential Revision: D1370893