Optimize ini_get
Summary: Add support to the JIT to recognize calls to ini_get for system settings and
substitute them with the value of the setting.

As part of this diff, I also removed the
system_get_custom_bool_settig API I added since it's no longer needed.

I also fixed an inconsistency between the version of IniSetting::Get
that takes a String& and a std::string& to return the setting value. The
std::string& version was incorrectly formatting double settings (they
included trailing 0's after the decimal point).

Reviewed By: @ptarjan

Differential Revision: D1378334