Elaborate namespace defs during parsing
Summary: Instead of having each caller to the parser need to remember to
call the namespace function, just do it in the parser. This is important
since it completely eliminates `Namespace` and `NamespaceUse` from the
AST, and basically everyone else will `assert false` if they see it.
(Notably, the JS typechecker was basically completely broken with
namespaces due to this.)

In an ideal world, these would be two different types and the OCaml
typechecker would yell at us for getting this wrong, but defining a
whole new AST for this seems like overkill.

Reviewed By: @elgenie

Differential Revision: D1369259