Update the behavior of NaN comparisons
Summary: The behavior of <= and >= in the presence of NaN changed sometime
between php 5.3 and 5.5. We were still giving and testing the 5.3
behavior, where e.g. ##NAN <= NAN## is true. (This is because it was
implemented as ##!(NAN > NAN)##.)

Change to the 5.5 behavior where all comparisons involving NaN, except
!= and !==, are false. This matches pretty much every other programming
language that has floating point numbers.

We're still not getting this perfectly right w.r.t. php5.5: ##"" < NAN##
should evaluate to true (I have not the faintest idea why) but we'll say
it's false.

Note that we currently don't jit any relational comparisons involving
doubles (even though we have codegen that's supposed to handle it), so
this change to tv-comparisons.cpp is all that's needed.

The new .expect file was generated with php5.5.

Reviewed By: @swtaarrs

Differential Revision: D1372193