Adapt naming to new error system
Summary: The new error system doesn't stop at the first error.
As a consequence, when an error occurs, an tree must be produced.
This diff introduces a new expression 'Any' that is produced when an error is encountered during the naming phase.
Any translates to the type Tany during type-checking, which should not introduce cascading errors.

Reviewed By: jwatzman, eletuchy

Differential Revision: D1372757