Adapt parser to new error system
Summary: The new error system in Hack doesn't throw anymore.
This diff adapts the parser to the new system.
The parser use to have a ~fail option that would throw when true and recover otherwise.
The recovering mode was mostly used for auto-completion.
This option doesn't make sense anymore (since errors don't throw and from now on, should never throw).
Instead, the parser returns the list of errors. The caller gets to decide what to do with that list of errors (it could ignore it).
I removed hh_read_file because Utils.cat is as efficient (probably because it uses a buffer).

Reviewed By: @joelbeales

Differential Revision: D1372768