Fixed property access to DOM and XMLReader classes (issue #1989)
Summary: Previously properties in the DOM extension and XMLReader class could be
accessed case-insensitively, because the properties were backed by a
case-insensitive map. This fixes that.

isset on DOM classes was broken before, this also fixes that.

The behavior of accessing invalid properties on DOM classes was also
inconsistent, this has been fixed. Before, unlike PHP and unlike other
objects, accessing an invalid property on DOM classes would result in
a fatal error.

Changing the type of map backing the DOM properties results in a different
ordering in var_dump. I've changed the tests to reflect this
(the ordering didn't match php before, so this doesn't break compatibility
there).

Reviewed By: ptarjan, liranuna

Differential Revision: D1362865