Stop storing two models in hh_ide.js
Summary: Motivation: As more and more code gets converted to hack, we're pulling more and more dependencies into hh_ide.js. The model is getting too big, and it's going to start crashing chrome for people. (This is started to happen occasionally to people now)

I tried adding deps like they were decl (as discussed with Josh when we were talking about this), but it only reduced the number of files added by 10%. :(

This should be a much bigger win. I think this code is basically legacy stuff from when we first integrated this. We were holding two models, one for typechecking and one for autocomplete. There doesn't seem to be much reason to do this now (especially now that @pikatchu is making us not stop on errors).

I'm hoping I'm not missing something here and that this is ok to do.

Seems like a reasonable assumption that holding one model instead of two will cut our declare_file time in half and the model size in half. This should buy us a bunch of time.

Reviewed By: @jwatzman

Differential Revision: D1374505