Band-aid apparently incorrect assertion in SimpleVariable::analyzeProgram
Summary: This is asserting something isn't a parameter when it doesn't
have various contexts flags, but it's not true if you didn't run with
PostOptimization and GenerateInferredTypes set, which are both
user-controlled flags.  I suspect the bug is the assert, not somewhere
else, but I don't know the code well enough to be sure.

Reviewed By: @markw65

Differential Revision: D1371177