Allow Repo::loadGlobalData to fail and print SrcKey's resumed flag
Summary: Print the SrcKey's resumed flag so we can more easily figure out why
we have multiple translations for the same md5 + bytecode offset.

Also, running hhvm -vEval.DumpTC=1 and then tc-print was failing
because there was no global data in the repo.  So allow
Repo::loadGlobalData to fail in this case.

Reviewed By: @swtaarrs

Differential Revision: D1372475