Fix /stop when bind fails, redux
Summary: Third attempt at fixing this problem.  The IP and AdminPassword are missing in the stop url.  This causes a 60 second delay starting a new HHVM as it waits to kill based on the pidfile.

Added a callback class 'ServerEventListener' which gets a call when a server is about to close the listening socket.  HttpServer uses this callback to remove the pidfile.  The timing of removing the pidfile is sensitive as we have to remove it before a new HHVM thinks it created it.

Fixing this now in the context of a mysterious startup crash that happens when devs get impatient and restart the server multiple times. t4483583

Reviewed By: @bertmaher

Differential Revision: D1377159